Add static_cast<int>() to resolve compiler warning (msvc x64).
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10129 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/core/include/mcrl2/core/dparser.h
@@ -223,7 +223,7 @@ struct parser
   {
     m_parser->start_state = start_symbol_index;
     m_parser->partial_parses = partial_parses ? 1 : 0;
-    D_ParseNode* result = dparse(m_parser, const_cast<char*>(text.c_str()), text.size());
+    D_ParseNode* result = dparse(m_parser, const_cast<char*>(text.c_str()), static_cast<int>(text.size()));
     if (!result || m_parser->syntax_errors)
     {
       throw mcrl2::runtime_error("syntax error");
